Code like it's 2015 - custom list perference
Create a custom list preference, using a RecyclerView to display the choices.

Add some padding to the top of the title, and to the bottom of the list items, to ensure that the scrollbar will always appear (at least on the different emulators tested).

This is in the hopes that the app will finally be approved.